tests/SwissTransportTest/TransportTest.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using System;

namespace SwissTransport
{
    [TestClass]
    public class TransportTest
    {
        private ITransport testee;

        [TestMethod]
        public void Locations()
        {
            testee = new Transport();
            var stations = testee.GetStations("Sursee,");

            Assert.AreEqual(10, stations.StationList.Count);
        }

        [TestMethod]
        public void StationBoard()
        {
            testee = new Transport();
            var stationBoard = testee.GetStationBoard("Sursee", "8502007");

            Assert.IsNotNull(stationBoard);
        }

        [TestMethod]
        public void Connections()
        {
            string date = "2020-3-24";
            DateTime time = Convert.ToDateTime(date);
            testee = new Transport();
            Connections connections = testee.GetConnections("Sursee", "Luzern",time);

            Assert.IsNotNull(connections);
        }

        
    }
}
